refactor: Inline yaml templates and fix bug (#23418)
## Description

The `upload-json-steps.yml` and `download-api-extractor-artifact.yml`
templates were not really that reusable so this PR inlines them into the
single pipeline that uses them.

It also fixes a bug that caused the JSON artifacts to never be published
as `latest-v<major>.tar.gz` because we were trying to use a variable
that only exists at runtime
(`dependencies.check_branch_version.outputs['SetShouldDeploy.shouldDeploy']`)
in a "compile-time" condition in `upload-json-steps.yml`. The condition
was thus never true, and the step under it was never present in the
pipeline. So I replaced a couple of compile-time conditionals with
`condition: ` nodes in the corresponding tasks, which should make the
tasks always appear in the pipeline, but get skipped unless the
specified conditions are true.