(tree): Support for adding constraints on revert to view.runTransaction API (#23400)
Added support to `view.runTransaction` API to allow transaction authors
to specify constraints that should apply to the revert of a changeset.
These constraints called `preconditionsOnRevert` should be returned by
the transaction callback. An alternative would have been to specify
these in the `runTransaction` params. However, transaction authors may
need to know the status of the tree when specifying the constraints. For
example, the constraint may be applied on a node that is created in the
transaction body and so, specifying them beforehand would not be
possible or rather, would not be straightforward.

These constraints will be checked and validated after the transaction
body has run and if they fail, the `runTransaction` API will throw a
usage error. Once the transaction is committed, the constraints will be
checked if and when the commit corresponding to the transaction is
reverted. If the constraints are violated at that time, the revert will
be ignored.
 

[AB#8069](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8069)