Abstract customdataRetriever in historian, pass simplified customData to gitrest filesystem (#23354)
## Description

This PR is to abstract customdataRetriever in historian and passing
simplifiedCustomData to gitrest.

Background:
Blob container per doc is a storage optimization for FRS Durable
Container. In container per doc mode, the document vs azure blob
container is 1:1 mapping. The blob container was created before the
initial summary uploaded.

To support a cmk-related case in containerPerDoc mode, there is specific
tenant-level info needs to be read from tenantDB, and passed to storage
layer in document creation flow. In oss, just wrap this info as a
generic message, and pass it all the way into filesystemFactory
parameters without editing, in this way most of implementation details
specific to this case can be contained in FRS.


Main Changes:

In historian, abstract an interface ISimplifiedCustomDataRetriever.get()
to convert customData to a simple string (customized implementation in
FRS), then pass to gitrest via http header.
In gitrest, get SimplifiedCustomData from header and add as part of
FileSystemManagerFactory parameter.