fix(property-changeset): Fix bug in IndexedCollection class (#23243)
### Description

We encountered an issue in the` fluid-experimental/property-changeset`
package within the `IndexedCollection` class. The Lodash `without`
method is being used with an Object as input, resulting in the following
error when applying a changeset:

`Error: PR-008: Trying to remove a non-existing entry:`

This issue occurs because `without` always returns undefined when used
with an Object, causing comparisons between the resulting values to
consistently evaluate as true. By replacing `without` with Lodash's
`omit`, which is designed for handling Object inputs, the issue is
resolved.

---------

Co-authored-by: Neeraj Charokar <neeraj.charokar@autodesk.com>
Co-authored-by: Tyler Butler <tylerbu@microsoft.com>