Make forest config tree-shake compatible (#23488)
## Description

Establish a new pattern for configuration to replace enums with
something compatible with tree shaking.

This has the annoyance that the different options are no longer grouped
under a scope.
It has the benefits that different values for the same option can now
have different release tags,
and that the code supporting the unused options can be removed with tree
shaking (including for users of the public API, which can't see the
alternative options but used to include them in the bundle anyway).

Apply this pattern to ForestType, giving a >7KB reduction in the tree
bundle, as measure by our bundle size tests.

## Breaking Changes

This is an incompatible change to the forest configuration alpha APIs:
users will need to point to the new opaque objects instead of enum
members. All impacted code will fail to build rather than having runtime
errors.