Simplified ScheduleManager and renamed it to InboundBatchAggregator (#23446)
Renamed `ScheduleManagerCore` to `InboundBatchAggregator` because its
primary objective is to ensure that an incoming batch is fully available
in the delta queue before it starts processing.

Also, simplfied it as following:
- Removed `ScheduleManager` because all it was doing was calling
`DeltaScheduler` and emitting `batchBegin` and `batchEnd` events. Moved
the event emission directly into container runtime which was recently
refactored to consolidate this logic in one place. `DeltaScheduler` now
processes batches on these events from container runtime.
- Removed the logic to modify `batch` metadata in delta manager's
`prepareSend` event handler. This is an artifact of very old logic where
adding the batch metadata was distributed between container runtime and
delta manager. For a long time now, container runtime is in full control
of adding (or not adding) batch metadata to messages in a batch. This
was done by https://github.com/microsoft/FluidFramework/pull/11832.
- Removed the `scheduleManager.spec.ts` unit tests. All it was doing was
validating `batchBegin` and `batchEnd` events which is the same as the
ones done in `batching.spec.ts` unit tests. Moreover, it was using
custom logic to identify batch beginning and end from the inbound
messages so it wasn't really testing any real code.

[AB#1981](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1981)