refactor(api-markdown-documenter): Consistent release levels (#23507)
Interfaces were previously introduced to represent parameters to HTML
rendering functions. They were added as `@public`, but the related
functions are `@alpha`. This PR makes the interfaces `@alpha` to align
with their corresponding functions.

While this is technically a breaking change, the types strictly exist
for input to functions that are `@alpha`, so this shouldn't have any
impact on users.