Change shares api path to match other use cases (#23434)
## Description

odsp-driver calls '/shares' API to redeem access to file referenced by
redeemable sharing link. I am changing API path to include '/driveItem'
in it. Technically it is not required for executing redeem operation.
However, we have other cases that use '/shares' API and do require to
specify '/driveItem' in order to get specific drive item properties. The
reason this matter is when caller of this API must possess logical
permissions to call this API (for instance, this will be the case when
call is made with app-only token) then two separate logical permissions
are needed for the '/shares' call with and without '/driveItem'.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

---------

Co-authored-by: Andrei Zenkovitch <andreize@microsoft.com>