Allow for passing client display name used when attributing changes to the current client in the collab session (#23422)
Allow for passing client display name used when attributing changes to
the current client in the collab session

## Description

This change introduces support for overriding the client display name
used when attributing changes to the current client in the collab
session. This can be used in scenarios where the client display name
should be different from the default client display name, for instance,
when client is running in app-only context.

This is achieved by adding a new optional property `displayName` to the
`ICollabSessionOptions` interface. This interface is passed inside
`HostStoragePolicy` which is specified when constructing document
service factory. If `displayName` is specified, then it is used when
establishing collab session and passed in the body of /joinSession
request to Push service.

---------

Co-authored-by: Andrei Zenkovitch <andreize@microsoft.com>