Remove _createDataStoreWithProps (#22996)
[AB#856](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/856)
https://github.com/microsoft/FluidFramework/issues/22993

### Description
Removes APIs that have been deprecated. We have tried to move away from
`_createDataStoreWithProps` as this causes collisions when creating live
datastores. Solutions involved aliasing. The original issue with the
removal was that there was no way to put in props. Added the
`initialState` parameter to
`PureDataObjectFactory.createInstanceWithDataStore` as pass in props.

### Remove APIs
- `IContainerRuntimeBase._createDataStoreWithProps`
- `ContainerRuntime._createDataStoreWithProps`

### Alternatives
- `PureDataObjectFactory.createInstanceWithDataStore` allows props to be
passed in via the `initialState` parameter.

### Merge Notes
- Merge only after `main` is ready for 2.20 changes.

---------

Co-authored-by: Tyler Butler <tylerbu@microsoft.com>