fix(BlobManager): Don't log getBlob errors if the runtime is disposed (#23521)
There's a race condition where the application may be closed while the
BlobManager is awaiting a blob to be downloaded from the service. When
this happens, the error "Storage Service is disposed. Cannot retry" is
thrown, and the BlobManager logs it as an error. But this isn't really
an interesting error (because we don't care about the result at this
point anyway) and shouldn't be logged as such.

This change downgrades the `BlobManager:AttachmentReadBlob_cancel` log
to "generic" category in this case, not "error".