improvement(fluid-build): Fix group task dependencies and reduce message clutter (#23523)
Fix group task dependencies: For script task that has "&&" in the
command, a sequential group task is created. When building the
dependency graph, the latter tasks need to depend on the previous task.
However, an "iterator" of previous leaf task is passed to add as
dependency for the latter tasks, and if the latter tasks have multiple
subtasks, the iterator will only iterate once for the first subtask and
miss the other subtask since it doesn't reset. The fix is to just pass
the Set into `addDependentLeafTasks`

Reduce message clutter: 
- For set up without config file, avoid having a big warning message.
Merge the indication with the "Build Root" message.
- Don't show symlink status unless it not in the default mode.

Additionally, make `getFluidBuildConfig` correctly track which
`searchDir` is using a default config in case it is used with multiple
`searchDir` in the future.

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>