refactor(runtime-definitions): Move to recommended lint config (#23555)
## Description

Moves the runtime-definitions package to use the recommended lint
configuration. Locally disabled rules for use of `any` in a few places,
and replaced one in a `@sealed` interface with `unknown`, additionally
updating the code that declares listeners for the relevant event.
Somehow TS doesn't complain about the mismatching listener signatures. I
did not try to investigate why.

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>