[LTS] refactor(ci): Remove scheduled run for e2e tests pipeline (#23601)
## Description

Removes the scheduled run of the e2e tests pipeline in the LTS branch.
Since we added the scheduled run of the build pipeline, we've been
getting 2 runs of the e2e tests one every day because it also triggers
in response to the completed build pipeline run. That one is enough for
our purposes.