(GC) Update a GC end-to-end test to use sinon timers instead of real ones (#23516)
This PR updates one of the GC end-to-end test to use sinon fake timers
instead of real timers. The tests in this file adds delays to wait for
sweep timers to expire. Using fake timers these delays have been
replaced with advancing the clock thereby reducing the runtime.
The run time for this test file went from 4s to 2s.


[AB#11248](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/11248)