improvement(build-common): Allow APIs to `{@link}` to APIs of lower release levels. (#23533)
Since we generate API reports for different release levels, it is
possible (and valid) for `@link` tags to reference an API item that
won't be visible at that level. Currently, our configurations forbid
such links as a side-effect of running link validation during report
generation (where APIs will be omitted from report consideration based
on their release tags.

However, it is often useful for APIs to link to more experimental APIs
for discoverability.

To make this possible, link validation is now disabled in report
generation configurations. It is maintained in model generation configs,
which is where we really need it.

Also adds a couple of such links in the `tree` package as a test (and
because we wanted them there).