build: Lockfile churn (#23626)
## Description

Changes that result from adding a dependency somewhere, doing `pnpm i
--no-frozen-lockfile`, removing the dependency, and doing `pnpm i
--no-frozen-lockfile` again.

Seems to be a lot of churn that the upgrade to pnpm v9 should have
applied but didn't. A couple breaking changes in [the v9 release
notes](https://github.com/pnpm/pnpm/releases/tag/v9.0.0) talk about peer
dendencies, which (part of?) this seems to be related to.