Parallel testing on client pipeline at job level (#23204)
This is an effort to reduce time consumed by our client pipeline,
parallelizing our tests running them on different agents. With this
change we're able to reduce the time from an average of 54min to 38min,
writing also the foundations to further parallelization that could keep
improving the time all the way down to 25 min on my estimations but
which work requires more than this wave of changes.

On a nutshell, this change creates a different template from which the
client pipeline runs, which main difference is the creation of an
artifact that contains some of the build output to be able to run tests
on different agents. After several optimizations, by publishing content
generated inside lib/dist directories and restore this content in a
different agent (once we have checked out the repo and install
dependencies), we're able to run our ci tests without needing to run the
build process again.

<img width="232" alt="image"
src="https://github.com/user-attachments/assets/9503bfde-c5ad-4676-8dfd-a7fcb8ae4a3b"
/>


Some caveats:
- Losing some code coverage: given the way coverage works, coverage
tests need to be run in the same agent. By running mocha and local e2e
tests in the same agent, we're able to keep coverage logic in place
altough we will lose coverage done by jest tests.
- jests tests need to run the build: for unknown reasons, jests tests
are the only ones failing with current approach of consuming lib/dist
directories. I even tried getting output.log and tsbuildinfo files but
yet there is something generated by the build necessary for these tests
to run, so for now we could also run c:build with these tests. It does
not matter in terms of time since tests don't consume a lot of time.
- repeated logic on our templates: new template shares a lot of logic
with build-npm-package.yml since it started as a copy of it. Also it has
a lot of logic that does not need to be there since we now know it will
only run for client. I don't consider these to be blocking current PR
but I would like to attend this in a future PR
- More agent creation: more agents are created in our pipelines given
this approach so I'll do some testing to make sure there is no
unexpected behavior when multiple pipeline runs at the same time.
- further parallelization: A lot of these caveats need to be solved to
further parallelize these tests. For example, we could split our e2e t9s
tests into 2, reducing significantly the time they consume. I believe we
could further reduce pipeline duration to 25-28min with further
investing.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>