Allow more generous waiting in attachLifecycle tests (#23646)
## Description

Adjusts the `timeoutPromise` calls in the attach lifecycle test suite to
use the default behavior of waiting until *just before* the test times
out rather than half the duration of the test. The previous behavior
effectively enforced that every operation completed within half the test
timeout, or 2.5s on tinylicious. That is indeed true most of the time,
but because we run so many variants of this test, it has been mildly
flaky on CI. After this change, we just enforce that the overall test
completes within 5s. There have been 6 hits recently for tinylicious,
with 4 of them timing out on a promise near the start of the test
duration. Here's a column chart of number of test runs vs duration in
ms:


![image](https://github.com/user-attachments/assets/d7ce6869-717d-42e6-8a08-efdcc8711cb0)

Interestingly, we have some passed runs over 3 second long. This
suggests some of the runs we're timing out because of one particular
step may not have even gone over 5 seconds.

The long tail here might mean we eventually want to also bump the t9s
timeout for this test suite, but we can evaluate that later using
similar techniques.

Co-authored-by: Abram Sanderson <absander@microsoft.com>