refactor(presence-tracker): Use presence package (#23020)
## Summary of changes

The presence-tracker example has been updated to use the presence
package. It now uses LatestValueManagers for mouse and focus tracking.

Each client also has a slider that controls the allowable latency of the
mouse updates. This enables quick demos of the differences in behavior
with higher latencies as well as demonstrating runtime adjustments of
the allowable latency.

## Known issues:

- The failing test checks that when a second client joins the first
client's client list is updated. This failure seems to indicate that the
multi-client setup is not working correctly. Follow-up issue:
[AB#28502](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/28502)