(compat) Added supported features and generation across the Loader / Runtime boundary (#22877)
## Description
Added `supportedFeatures` and `generation`properties to Runtime and
Loader layer boundary. These will be used to validate that the layers
are compatible.
These two properties along with `pkgVersion` are added to a
`ILayerCompatDetails` interface. `ContainterContext` and
`ContainerRuntime` implements this interface provider and uses the
provider pattern to access it from the other object. The `interface` is
internal and is not exposed since this in an internal implementation.
`pkgVersion` is includes in the error / warning logs to give additional
context to users as to what the incompatible version combination is.
Also, added a `ILayerCompatSupportRequirements` interface that has
properties that a layer requires other layer to be at to be compatible.

### Supported features
- Both the Runtime and Loader layers have a `supportedFeatures` property
that includes a set of features that a layer supports. This is
advertised to the other layer at the layer boundary.
- Both the layers have a `requiredFeatures` array that includes a set of
features that it requires the other layer to support in order to be
compatible. _Note that this is internal to that layer and is not
advertised._
- At runtime, each layer validates that all the entries in
`requiredFeatures` is present in the `supportedFeatures` of the other
layer. If not, it closes the container with a usage error that says
layers are not compatible.
- Any change that adds new behavior or changes existing behavior across
a layer boundary must add an entry to the supported features set. For
example, changes such as "move writing protocol tree to summary from
Loader to Runtime" should add an entry to the supported features set.

### Generation
- In addition to supported features, a `generation` is also added to
both the layers. This is advertised to the other layer at the layer
boundary.
- Generation is an integer which will be incremented every month. It
will be used to validate compatibility more strictly (time-based) than
supported features where layers are incompatible only if there are
unsupported features.
_Note: The logic to update the generation will be added in a later PR.
See
[AB#27054](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/27054)
for a proposed solution._
- One key reason for adding generation is to have a clear cut off point
where we officially stop supporting layer combinations. Our compat tests
will test combinations up to this cut off point. The idea is that we
test what we support and whatever we don't test should not be supported.
This will help us achieve that.
- Say that Runtime / Loader compatibility is N months. We can start
throwing warnings when the layers are almost N-months apart saying that
this combination is not supported. This will give applications an early
signal that layers are about to be incompatible and they can take
necessary steps. We can also choose to close the container if we decide
to be strict about the layer combinations.
- Both the layers have a `minSupportedGeneration` that it needs the
other layer to be at. _Note that this is internal to that layer and is
not advertised._
- At runtime, each layer validates that the `generation` of the other
layer is >= `minSupportedGeneration`.
- For example, say that the Runtime is at generation 20 and has a
`minSupportedGeneration` of 8 for the Loader layer (12 month support
window). If it sees that Loader's generation is lower than 8, the layers
are incompatible.

### Backwards compatibility
To support older layers before layer compat enforcement is introduced,
the `minSupportedGeneration` is set to 0 whereas `generation` starts at
1. For older layers that did not implement `ILayerCompatDetails`, their
generation is set to 0 and supported features is set to an empty set.
So, their generation is compatible until we update the
`minSupportedGeneration`.


[AB#20805](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/20805)