fix(test-version-utils): Fix version calculation to account for legacy breaking minor releases (#23603)
This PR fixes N-X version calculation to take into account the new
legacy breaking minor releases. For example,

- For `2.0.0 <= N < 2.10.0`, N-1 is 2.0.0-rc.5.0.x.
- For `2.10.0 <= N < 2.20.0`, N-1 is 2.5.x (latest minor between 2.0 and
2.10 is 2.5.x)
- For `2.20.0 <= N < 2.30.0`, N-1 is 2.13.x (latest minor between 2.10
and 2.20 is 2.13.x)

We should still consider a full rewrite of version calculation, since
the code is extremely messy and hard to follow for developers not
familiar with it already. However, this change is needed to ensure we
continue testing properly in the meantime. See
[AB#8198](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8198)
for more details.


[AB#28437](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/28437)