Fix and reorganize stashed ops that wait for a summary (#23542)
Some tests among stashedOps.spec.ts need to wait for a summary in order
to fall into its test scenario. There was some missing configuration on
them:
- ensureSynchonized should be called before summarizing to ensure that
it summaries all changes up to that point.
- When created the provider via getTestObjectProvider, pass in option {
syncSummarizer: true } so that ensureSynchonized will sync summarizer
clients.
- explicitly load summarizers from the snapshot generated by previous
summarizers by specifying the version of that snapshot

Besides those changes, I made some reorganization to these tests since
added config is necessary only for tests that wait for a summary so I
separated them into a different file and move some functions to a utils
file well.

Some of these tests are currently failing in ODSP and so I'll head up
OCE in case they keep failing after this PR is merged. If this approach
work, I'll use it for refresh snapshot tests that are currently failing
for what I suspect the same reason.