Reapply 'Use frozen lockfile in routerlicious build' (#23656)
## Description

Reapplies #23558 with a fix for the code coverage issue. That issue
ended up being that the `build-docker-service.yml` template takes a
dependency on the package location in the base image (*not the runner
image*), which that PR modified. This didn't come up for other server
builds as they don't report coverage.

Reviewer is encouraged to look by commit.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>