Surface internal error codes correctly (#23681)
## Description

Previously, handleResponse() would override internal error codes with a
default status of 500. This change ensures that we only fall back to 500
when no valid internal error code is present. This specifically impacts
the getDeltas API, where certain cases incorrectly returned 500 instead
of 404.

> For bug fixes, also include specifics of how to reproduce it / confirm
it is fixed.

Added tests to reproduce the issue and verify the fix

## Breaking Changes

For the cases where getDelta() was called on a non-existent document, we
now return 404 instead of 500

---------

Co-authored-by: Pradeep Vairamani <pradeep@Pradeeps-MacBook-Pro-2.local>