Remove support for Batch Compression without Batch Grouping (#23017)
## Description
We can and should get rid of the mode where we send ops with empty
content for Compression.

Configuration-wises, at this point both compression & op grouping are ON
in main, and OFF in LTS. So, we can group them together into one
modality and stop mentioning (and dealing with) a modality where
compression is on, but op grouping is off.

Acceptance Criteria

The code that compresses batches of 2 or more messages by creating empty
message placeholders should be removed. The code to reconstitute such
batches must remain, of course.

Execution Plan

There are two exceptions we need to deal with first as part of this:
* Stop compressing the blobAttach batch, since Grouping is disabled (See
discussion in ADO for justification of this change)
* Always group a batch of 2 or more messages - Remove the ability to
configure this

Then it's just a matter of updating OpCompressor to require
single-message batches and remove the empty placeholder stuff.

## Reviewer Guidance

Let me know if there's a better way to accomplish the task. Also review
the tests to make sure that the flow the tests are trying to recreate is
still being followed.

Fixes:
[AB#28649](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/28649)