fix(build-cli): Only update range for workspace:*|^|~ dependencies (#23669)
## Description

Makes it so `flub bump` only updates the range for `workspace:*|^|~`
dependencies. The problem is that even when we have dependencies like
`"fluid-framework": "^1.4.0"` in and end-to-end tests package, which is
supposed to point to FF 1.x, it currently gets replaced with an exact
dependency on `X.Y.Z-<buildId>` and this breaks some test pipelines.