refactor(ci): Disable CodeQL the right way (#23691)
## Description

Fixes the way we disable CodeQL for the test jobs in the build - client
pipeline, so they actually cause the tasks to not trigger (what we had
done applies when the CodeQL tasks are added manually, but when they are
added by the 1ES pipeline templates we needed a different approach).

Also, it disables CodeQL tasks for all non-main branches. Turns out that
the default logic in 1ES templates for "only run CodeQL for the default
branch in the repo" only works for repos hosted in ADO. For us, that
resulted in all branches triggering CodeQL tasks, which unnecessarily
add a significant amount of time to pipeline runs when they trigger.