build(docs): Don't generate API documentation for `@fluid-experimental` and `@fluid-example` packages (#23690)
We don't intend for our public customers to use these packages, and none
of our production libraries' APIs reference their contents. This removes
API docs for the following packages:
- @fluid-example/example-utils	
- @fluid-example/migration-tools
- @fluid-experimental/attributable-map
- @fluid-experimental/attributor	
- @fluid-experimental/data-object-base
- @fluid-experimental/data-objects
- @fluid-experimental/dds-interceptions	
- @fluid-experimental/ink	
- @fluid-experimental/last-edited	
- @fluid-experimental/oldest-client-observer	
- @fluid-experimental/ot	
- @fluid-experimental/pact-map	
- @fluid-experimental/property-dds	
- @fluid-experimental/sequence-deprecated	
- @fluid-experimental/sharejs-json1	
- @fluid-experimental/tree-react-api

This policy can be reverted at any time if we wish to restore some/all
of these to the website. For now, it reduces clutter and build times.