test(client): remove mismatched fluid-framework dep (#23682)
azure-client 1.x declares a peerDependency on fluid-framework 1.4.0.
This should not be elevated as normal dependency in end-to-end-tests and
that dependency is removed. (Note that pnpm will allow local workspace
version to satisfy that `^1.4.0` requirement, which is also a concerning
pattern.)
Since there is no actual peer requirement (it should have been an
optional peer dependency), just remove the peer dependency when
resolving dependencies using pnpm hook (`.pnpmfile.cjs`).
Then also remove the end-to-end test dependency that had been added to
satisfy the peer dependency. (May come back as `fluid-framework-legacy`
in later change.)