Replace IsomorphicPerformance with a performanceNow() method (#23668)
IsomorphicPerformance was an internal API in client-utils. It included a number of methods supported by the [Performance](https://developer.mozilla.org/en-US/docs/Web/API/Performance) API, including `mark()` and `measure()`. The only member of the API that was used as of this change is `now()`. This change removes IsomorphicPerformance, exposes `performanceNow()`, and reduces the amount of work needed to restructure client-utils.

Note that there is another copy of IsomorphicPerformance in common-utils. The common-utils version is deprecated, so this PR does not change it.