fix(client-presence): LatestValueManager with array comes across as object (#23662)
## Bug Description
[ADO Bug
29110](https://dev.azure.com/fluidframework/internal/_workitems/edit/29110)

Using an array (at least using `string[]`) as the type for `Latest`
(`LatestValueManager`) appears to rehydrate as object rather than array.
Example result:
```JSON
{
   "0": "foo"
}
```

Repro branch:
https://github.com/jason-ha/FluidExamples/tree/presence_Latest_array
Repro: connect two clients; make selection in one
Expected: selection shows in other
Actual: runtime error `indexOf` is not a member of ...

## Solution
When we create a LatestValueManager with an`initialValue`, we make a
shallow copy by doing `{ ..initialValue}`. So if `initialValue` is an
array we're transforming it into an object. To avoid this we can check
if it's an array first and maintain array.
Was able to solve the repro with this fix.