Export SharedTree DDS type and attributes to be used in Loop Repo for creating a Delay Loaded Factory (#23731)
## Description

Export SharedTree DDS type and attributes to be used in Loop Repo for
creating a Delay Loaded Factory. We needed to created a delay loaded
factory for shared tree in Loop Repo in to not impact the bundle size on
load. So, both the type and attributes needs to be exported.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>
Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>