Export Map and Directory Fuzz Models (#23754)
This change follows the same pattern used for merge tree to export test
only utilities from the map package. The utilities that are exported are
the base fuzz testing model for map and directory. Exporting the base
model will allow their reuse in other testing contexts. In order to
export these models, they have to be put into another file,
fuzzUtils.ts, as exporting from a test file (`*.spec.ts`) results in the
test running in that file re-running where anything is imported from it,
which is not something we want.

The in-progress PR shows the expected usages of these imports:
https://github.com/microsoft/FluidFramework/pull/23701