remove(devtools): Prototype editing code (#23741)
This code was developed as a prototype but was never formalized and is
not on the near-term product roadmap. Removing the code for now to
reduce maintenance burden.

Note: backwards compatibility with the browser extension is not a
concern here for 2 reasons:
1. The API exposes no way to enable editing, so no clients could be
using the functionality
2. Even if somehow the browser extension was sending edit messages to
the client, they would simply be ignored.