fix(server): Fix builds in response to recent corepack issue (#23769)
## Description

Builds of server components, which happen inside Docker containers, were
broken by a recent issue with corepack + the npm registry. The registry
rotated some keys used for package signing, but corepack versions have
hardcoded references to the old keys. The latest pnpm packages are
signed with the new keys, so when corepack tries to download them it
fails.

This PR fixes the problem by making sure we install the latest (fixed)
version of corepack before using it.

See: 

- https://github.com/nodejs/corepack/issues/612
- https://github.com/pnpm/pnpm/issues/9029