refactor(aqueduct, fluid-static): Refactor internal "object class" types (#23784)
Renames "DataObjectClass" and related API members to "DataObjectKind" to
better match related APIs (e.g. "SharedObjectKind").

Also:
- ~~Moves `DataObjectKind` from `fluid-static` to `aqueduct`~~
- Reverted due to mysterious downstream build failures. Will attempt to
do this in a follow-up PR.
- ~~Update `createDataObjectKind` function in `aqueduct` to leverage
`DataObjectKind` directly.~~
- Reverted due to mysterious downstream build failures. Will attempt to
do this in a follow-up PR.
- Renames `LoadableObjectClass` to `LoadableObjectKind`

This PR is a small part of a larger effort to work towards supporting
`SharedTree`-backed `DataObject`s (as opposed to `SharedDirectory`-based
ones).