Reupload blob if expired (#23762)
We could be more resilient to blob expiration which could happen on
resubmitting blob attach ops. Instead of closing the container, we now
can reupload the same blob and kickstart the same process we had before.

This is attending a Vlad comment in which he suggested "we have to
ensure we restart upload process if it happens that blob upload and op
telling storage about it are far apart in time (above allowed limit by
storage). This can happen if we disconnect after uploading blobs."