(tree) Fixes a bug where trunk is not correctly trimmed with grouped batches (#23781)
## Bug
With grouped batching enabled, the trunk is not correctly trimmed if
there are multiple shared tree messages in the same batch. This PR adds
a test to demonstrate that scenario where there are 2 ops in a batch -
one to initialize the shared tree data and the other to remove the first
node. After the minimumSequenceNumber moves past the sequence number of
the above batched op, the trunk should be trimmed to remove the first
node, but it does not.

## Root-cause
With grouped batching enabled, the ops in a batch have the same
`sequenceNumber` and `minimumSequenceNumber`. Shared tree distinguihes
and orders ops in a grouped batch via a `SequenceId` which is of the
form `{ sequenceNumber: number, indexInBatch?: number }`. Note that if
`indexInBatch` is undefined, it indicates the first message in the batch
which is equivalent to `indexInBatch: 0`.

When`EditManager::trimTrunk` function is called at the end of op
processing to trim the trunk (if applicable), it gets the sequenceId of
the op / commit preceding the minimum commit of its trunk branches via
`decrementId` function. However, this function has a bug where it does
not take `indexInBatch` into account if the passed in `sequenceId` has
`indexInBatch` undefined. So, for sequence id of the format `{
sequenceNumber: N }` it always returns `{ sequenceNumber: N - 1 }` which
is incorrect if there were multiple messages in the batch with
sequenceNumber N-1.

## Fix
The `decrementId` function has been renamed to
`getUpperBoundOfPreviousSequenceId` to better reflect what it's doing
now. If the `sequenceId` being passed is the first message in a batch,
i.e., `{ sequenceNumber: N, `indexInBatch`: 0 or undefined }`, it
returns `{ sequenceNumber: N - 1, `indexInBatch`:
Number.PositiveInfinity }` indicating that the preceding op / commit can
have a maximum sequence number of infinity.
It also updates the `sequenceIdComparator` to take `indexInBatch`
infinity into account.