Update odsp-driver package tests to build and run in ESM (#23728)
## Description

Updates the package test scripts to build and run in ESM in addition to
CJS.

This required updating a few places where functions were being mocked by
the [sinon](https://sinonjs.org/) library.

The sinon `stub()` function mocks a given function on a given object.
For top-level functions, this meant that the exporting module object's
reference to the function was being replaced with the stub. This is no
longer permitted in ES6 since module objects are _readonly_.

To solve this, there is now a helper function `mockify()` which can
decorate a function with a property that allows it to be easily mocked
by sinon. The helper works by wrapping the function in a new function
with the same signature; the inner function delegates to a property that
has the actual function implementation (which defaults to the original
function). This pattern makes it easy to update this package without
making significant changes to the code flow. Rather, the problematic
functions and their mock sites have been updated to use this helper. One
function, `fetchHelper` was already a wrapper around `fetch` (which was
being mocked) - so it has been updated to use a similar pattern but in a
slightly different way.

Also replaces two usages of the CJS-supported `__dirname` with the
`_dirname` pattern (see dirname.cts) to accommodate relative file paths
in both CJS and ESM.